Update cic-in-k8s.md
This guide has components hard-coded to the default namespace. I just added a statement on line 12 (first line where user will execute code) stating this guide requires use of the default namespace. I tried the lab in a different namespace and permissions fail because the ClusterRole bindings are in default namespace based on cpx.yaml